# Faster operand stack