* Uses a pre-allocated array.
* Avoids frequent growing and shrinking of array
* Convenience methods for `dup*` opcodes to avoid frequent pops and pushes

## Benchmark results
I tested with [two benchmarks](https://gist.github.com/hrj/94784fc6af4b49a112a0). `Compute` just computes some fibonacci numbers over and over, and the second one compiles a `.java` file using `javax.tools`

### Compute
|Browser|Before|After|Improvement|
|-------|-----|-----|-----|
|Firefox 43|2075ms|1105ms| 46%|
|Chromium 47|1049ms|774ms|26%|

### Compile
|Browser|Before|After|Improvement|
|-------|-----|-----|-----|
|Firefox 43|40215ms|30798ms| 23%|
|Chromium 47|26811ms|25197ms|6%|