@jvilk 2016-01-26
:clap: Awesome! The code looks great, too.

You can actually figure out how much stack space Java methods require. The `Code` attribute has the properties `maxStack` and `maxLocals` that indicate the needed stack/locals size. 

Would you be willing to hook up `maxStack` and repeat your experiments?

If not, I can merge this as-is.