@hrj 2016-01-27
Thanks! I have used `maxStack` and pushed the commit. I don't see any noticeable difference in benchmarks though.

Why did you mention `maxLocals`? I am not very familiar with this stuff hence not sure if I can use it in this PR.