@jvilk 2016-01-27
Typical jvilk over-explaining. My brain is a bit fried from writing up student assignments. It's not relevant. 

It's likely that the `maxStack` change only changes memory consumption a bit, since most JVM methods have a very short operand stack.

Thanks again for this! I'll merge it in.